refactor: StyledLink 컴포넌트 GlobalStyle로 정의
