--- app/src/utils/StyledLink.jsx
@@ -1,19 +1,0 @@
-import { Link } from 'react-router-dom';
-import styled from 'styled-components';
-
-const StyledLink = styled(Link)`
-  display: block;
-  width: 100%;
-  height: 100%;
-  color: ${({ textcolor }) => textcolor};
-  text-decoration: none;
-  &:focus,
-  &:hover,
-  &:visited,
-  &:link,
-  &:active {
-    text-decoration: none;
-  }
-`;
-
-export default StyledLink;
